requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
aiosmtplib==2.0.2
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
Authlib==1.3.1
bcrypt==4.1.3
black==24.4.2
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
distro==1.9.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.0
fastapi==0.111.1
fastapi-cli==0.0.4
fastapi-mail==1.4.1
frozenlist==1.4.1
google-auth==2.31.0
greenlet==3.0.3
h11==0.14.0
httpcore==0.17.3
httptools==0.6.1
httpx==0.24.0
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
mailjet-rest==1.3.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
openai==1.30.0
orjson==3.9.14
packaging==24.1
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.2.2
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic-extra-types==2.9.0
pydantic-settings==2.4.0
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==0.21.0
python-http-client==3.3.7
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
rsa==4.9
sendgrid==6.11.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.30
starkbank-ecdsa==2.2.0
starlette==0.37.2
tqdm==4.66.4
twilio==9.2.3
typer==0.12.3
typing_extensions==4.12.2
ujson==5.9.0
urllib3==2.2.1
uuid7==0.1.0
uvicorn==0.18.3
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
